Stabilise full upload→query loop; get KPR running; fix search UI honesty
Pipeline reliability (torch 2.8 / py3.9 / CPU):
- YOLOv8 weights_only=False shim in detect.py + track.py (torch>=2.6 rejects
  Ultralytics checkpoints by default)
- Add missing lapx dep for BoT-SORT (backend + ai_pipeline requirements)
- from __future__ import annotations in crop_extractor.py/embed.py (3.9 compat)

KPR part matching now runs locally:
- Gallery + query KPR moved to fresh-subprocess CLI calls (torchreid fork
  cannot coexist in-process with standard torchreid; its sys.exit also
  escaped except-Exception and froze upload jobs at step 5/5)
- Catch (Exception, SystemExit) on optional SOLIDER/face/KPR steps + query
  session handler so failures are non-fatal warnings
- Correct probed KPR dims in config.yaml (8 parts x 512-d)
- Query-time YOLO auto-crop: full-frame snaps crop to largest detected
  person instead of scoring 0

Frontend (app.js): tab-aware search submit; no enrolled targets / submit
errors now show real errors instead of the canned 94.2%% demo animation.
Add KPR test config, SOLIDER embed script, phase-5.8 search doc, README
changelog; ignore KPR/SOLIDER clones + OSNet gallery backups.

+++ ai_pipeline/reid/embed_kpr.py
@@ -481,7 +481,9 @@ def probe_kpr_output(
         part_dim  = embs.shape[2]
 
         # Holistic slot = highest mean visibility (foreground branch ≈ 1.0)
-        mean_vis = vis[0].detach().cpu()  # [num_slots]
+        # NOTE: vis can be a Bool tensor on some builds — argmax is not
+        # implemented for Bool on CPU, so cast to float first.
+        mean_vis = vis[0].detach().cpu().float()  # [num_slots]
         holistic_idx = int(mean_vis.argmax().item())
 
         num_parts = num_slots - 1
